fix(ci): file line limit — split oversized comment verdict coverage test
