+++ tests/unit/runtime/test_comment_verdict_coverage_edges_part_002.py
@@ -1,0 +1,3 @@
+"""Compatibility import target for split comment verdict coverage-edge tests."""
+
+from tests.unit.runtime.test_comment_verdict_coverage_edges_parts.test_comment_verdict_coverage_edges_part_002 import *  # noqa: F403
