+++ tests/unit/runtime/test_comment_verdict_coverage_edges_parts/_helpers.py
@@ -1,0 +1,78 @@
+"""Shared git worktree helpers for comment verdict coverage-edge tests."""
+
+from __future__ import annotations
+
+import subprocess
+from pathlib import Path
+
+
+def init_git_worktree(worktree: Path) -> None:
+    subprocess.run(["git", "init"], cwd=worktree, check=True, capture_output=True)
+    subprocess.run(
+        ["git", "config", "user.email", "test@example.com"],
+        cwd=worktree,
+        check=True,
+        capture_output=True,
+    )
+    subprocess.run(
+        ["git", "config", "user.name", "Test"],
+        cwd=worktree,
+        check=True,
+        capture_output=True,
+    )
+    (worktree / "src").mkdir()
+    target = worktree / "src" / "x.py"
+    target.write_text("base\n", encoding="utf-8")
+    subprocess.run(["git", "add", "src/x.py"], cwd=worktree, check=True, capture_output=True)
+    subprocess.run(["git", "commit", "-m", "init"], cwd=worktree, check=True, capture_output=True)
+
+
+def init_git_worktree_with_dirty_submodule(worktree: Path, *, submodule_name: str = "sub") -> None:
+    """Parent repo with a tracked submodule whose checked-out HEAD differs from the index."""
+    subprocess.run(["git", "init"], cwd=worktree, check=True, capture_output=True)
+    subprocess.run(
+        ["git", "config", "user.email", "test@example.com"],
+        cwd=worktree,
+        check=True,
+        capture_output=True,
+    )
+    subprocess.run(
+        ["git", "config", "user.name", "Test"],
+        cwd=worktree,
+        check=True,
+        capture_output=True,
+    )
+    submodule = worktree / submodule_name
+    submodule.mkdir()
+    subprocess.run(["git", "init"], cwd=submodule, check=True, capture_output=True)
+    subprocess.run(
+        ["git", "config", "user.email", "test@example.com"],
+        cwd=submodule,
+        check=True,
+        capture_output=True,
+    )
+    subprocess.run(
+        ["git", "config", "user.name", "Test"],
+        cwd=submodule,
+        check=True,
+        capture_output=True,
+    )
+    (submodule / "file.txt").write_text("v1\n", encoding="utf-8")
+    subprocess.run(["git", "add", "file.txt"], cwd=submodule, check=True, capture_output=True)
+    subprocess.run(
+        ["git", "commit", "-m", "sub init"], cwd=submodule, check=True, capture_output=True
+    )
+    subprocess.run(
+        ["git", "submodule", "add", f"./{submodule_name}", submodule_name],
+        cwd=worktree,
+        check=True,
+        capture_output=True,
+    )
+    subprocess.run(
+        ["git", "commit", "-m", "add sub"], cwd=worktree, check=True, capture_output=True
+    )
+    (submodule / "file.txt").write_text("v2\n", encoding="utf-8")
+    subprocess.run(["git", "add", "file.txt"], cwd=submodule, check=True, capture_output=True)
+    subprocess.run(
+        ["git", "commit", "-m", "sub v2"], cwd=submodule, check=True, capture_output=True
+    )
